+++ src/main/resources/templates/include/header.html
@@ -25,7 +25,7 @@
                     </th:block>
                     <th:block sec:authorize="isAuthenticated()">
                         <a th:href="@{/board/list}" class="btn btn-main">게시판</a>
-                        <a th:href="@{/member/mypage}" class="btn btn-outline-main ms-2 ms-lg-3">마이페이지</a>
+                        <a th:href="@{/mypage/info}" class="btn btn-outline-main ms-2 ms-lg-3">마이페이지</a>
                         <a th:href="@{/member/logout}" class="btn btn-main ms-2 ms-lg-3">로그아웃</a>
                     </th:block>
                 </div>
