+++ src/main/resources/templates/mypage/info.html
@@ -18,7 +18,7 @@
                                 <h4 class="widget-title text-center"><span>마이페이지</span></h4>
                                 <ul class="list-unstyled widget-list">
                                     <li><a class="" href="#!">대시보드</a></li>
-                                    <li><a class="" th:href="@{/member/mypage}">회원정보 수정</a></li>
+                                    <li><a class="" th:href="@{/mypage/info}">회원정보 수정</a></li>
                                 </ul>
                             </div>
                         </div>
@@ -57,7 +57,7 @@
                                         <th class="text-center pwd">비밀번호</th>
                                         <td>
                                             <!--<button type="button" class="btn btn-main btn-sm" onclick="reco()">변경</button>-->
-                                            <a th:href="@{/member/changePw}" class="btn btn-main btn-sm">변경</a>
+                                            <a th:href="@{/mypage/changePw}" class="btn btn-main btn-sm">변경</a>
                                         </td>
                                         <input type="hidden" name="pw" id="pw" class="form-control"
                                                th:value="${member.pw}">
@@ -99,7 +99,7 @@
                             <div class="text-center my-4">
                                 <button type="submit" class="btn btn-outline-main mx-1">수정</button>
                                 <!--<button type="button" class="btn btn-outline-main mx-1">취소</button>-->
-                                <a th:href="@{/member/mypage}" class="btn btn-outline-main mx-1">취소</a>
+                                <a th:href="@{/mypage/info}" class="btn btn-outline-main mx-1">취소</a>
                             </div>
                             </form>
 
@@ -122,7 +122,7 @@
 
                                     function leave() {
                                         if (confirm("탈퇴하시겠습니까?")) {
-                                            location.href = realpath + "member/remove?id=" + [[${member.id}]];
+                                            location.href = realpath + "mypage/remove?id=" + [[${member.id}]];
                                         }
                                     }
                                 </script>
